Fix broken project detail routes
