--- .env.example
@@ -1,3 +1,0 @@
-# Base URL of the OSK backend API (no trailing slash).
-# Copy this file to .env and adjust if you run the API locally.
-VITE_API_BASE_URL=https://api.oskigali.org/api
