+++ src/App.tsx
@@ -14,32 +14,34 @@ import Event from "./pages/event/Event";
 const PartnersForm = lazy(() => import("./pages/PartnersForm"));
 
 const wrap = (Component: ComponentType) => (
-  <Suspense fallback={<div>Loading...</div>}>
-    <Component />
-  </Suspense>
+	<Suspense fallback={<div>Loading...</div>}>
+		<Component />
+	</Suspense>
 );
 import ErrorPage from "./pages/ErrorPage";
+import ProjectDetails from "./pages/project-details/ProjectDetails ";
 
 const router = createBrowserRouter([
-  {
-    path: "/",
-    Component: RootLayer,
-    ErrorBoundary: ErrorPage,
-    children: [
-      { index: true, Component: HomePage },
-      { path: "/about", Component: About },
-      { path: "/community", Component: Community },
-      { path: "/event", Component: Event },
-      // {path:'/resources', Component: Resources},
-      { path: "/projects", Component: Project },
-      { path: "/partners", Component: Partners },
-      { path: "/partnersform", element: wrap(PartnersForm) },
-    ],
-  },
+	{
+		path: "/",
+		Component: RootLayer,
+		ErrorBoundary: ErrorPage,
+		children: [
+			{ index: true, Component: HomePage },
+			{ path: "/about", Component: About },
+			{ path: "/community", Component: Community },
+			{ path: "/event", Component: Event },
+			// {path:'/resources', Component: Resources},
+			{ path: "/projects", Component: Project },
+			{ path: "/projects/:slug", Component: ProjectDetails },
+			{ path: "/partners", Component: Partners },
+			{ path: "/partnersform", element: wrap(PartnersForm) },
+		],
+	},
 ]);
 
 const App = () => {
-  return <RouterProvider router={router} />;
+	return <RouterProvider router={router} />;
 };
 
 export default App;
