+++ src/hooks/useProjects.ts
@@ -3,35 +3,35 @@ import type { Projects } from "@/types";
 import { fetchProjects } from "@/api";
 
 interface UseProjectsReturn {
-  projects: Projects[];
-  loading: boolean;
-  error: string | null;
+	projects: Projects[];
+	loading: boolean;
+	error: string | null;
 }
 
 export function useProjects(): UseProjectsReturn {
-  const [projects, setProjects] = useState<Projects[]>([]);
-  const [loading, setLoading] = useState(true);
-  const [error, setError] = useState<string | null>(null);
+	const [projects, setProjects] = useState<Projects[]>([]);
+	const [loading, setLoading] = useState(true);
+	const [error, setError] = useState<string | null>(null);
 
-  useEffect(() => {
-    let cancelled = false;
-    fetchProjects()
-      .then((data) => {
-        if (!cancelled) {
-          setProjects(data);
-          setError(null);
-        }
-      })
-      .catch((err: Error) => {
-        if (!cancelled) setError(err.message);
-      })
-      .finally(() => {
-        if (!cancelled) setLoading(false);
-      });
-    return () => {
-      cancelled = true;
-    };
-  }, []);
+	useEffect(() => {
+		let cancelled = false;
+		fetchProjects()
+			.then((data) => {
+				if (!cancelled) {
+					setProjects(data);
+					setError(null);
+				}
+			})
+			.catch((err: Error) => {
+				if (!cancelled) setError(err.message);
+			})
+			.finally(() => {
+				if (!cancelled) setLoading(false);
+			});
+		return () => {
+			cancelled = true;
+		};
+	}, []);
 
-  return { projects, loading, error };
+	return { projects, loading, error };
 }
